+++ Scripts/Game/SecondaryBullet.cs
@@ -32,7 +32,7 @@ namespace Game
             if (elapsedTime >= ProjectileLifetime)
             {
                 Log("SecondaryBullet lifetime expired, destroying...");
-                InternalCalls.Scene_DestroyEntity((uint)EntityID);
+                InternalCalls.Scene_DestroyEntity((ulong)EntityID);
                 return;
             }
 
@@ -46,11 +46,11 @@ namespace Game
 
             for (int i = 0; i < collisionCount; i++)
             {
-                InternalCalls.Physics_GetCollisionPair(i, out uint entityA, out uint entityB);
+                InternalCalls.Physics_GetCollisionPair(i, out ulong entityA, out ulong entityB);
 
                 if (entityA == EntityID || entityB == EntityID)
                 {
-                    uint otherEntity = (entityA == EntityID) ? entityB : entityA;
+                    ulong otherEntity = (entityA == EntityID) ? entityB : entityA;
                     OnHitEnemy(otherEntity);
                     return;
                 }
@@ -65,7 +65,7 @@ namespace Game
             SpawnVortexAOE();
 
             // Destroy this bullet
-            InternalCalls.Scene_DestroyEntity((uint)EntityID);
+            InternalCalls.Scene_DestroyEntity((ulong)EntityID);
         }
 
         private void SpawnVortexAOE()
@@ -74,12 +74,12 @@ namespace Game
             {
                 // Get current bullet position before destroying
                 Vector3 currentPosition;
-                InternalCalls.Transform_GetPosition((uint)EntityID, out currentPosition);
+                InternalCalls.Transform_GetPosition((ulong)EntityID, out currentPosition);
 
                 Log("Spawning Vortex at: " + currentPosition.X + ", " + currentPosition.Y + ", " + currentPosition.Z);
 
                 // Instantiate vortex prefab
-                uint vortexEntityID = InternalCalls.Prefab_Instantiate(VortexPrefabPath);
+                ulong vortexEntityID = InternalCalls.Prefab_Instantiate(VortexPrefabPath);
 
                 if (vortexEntityID != 0)
                 {
